added where clause baseline
