+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Query/Parser/OrderByClauseSqlParserBaselineTests.cs
@@ -8,29 +8,47 @@ namespace Microsoft.Azure.Cosmos.Tests.Query.Parser
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     [TestClass]
-    public sealed class OffsetLimitClauseSqlParserBaselineTests : SqlParserBaselineTests
+    public sealed class OrderByClauseSqlParserBaselineTests : SqlParserBaselineTests
     {
         [TestMethod]
-        public void Tests()
+        public void SingleOrderBy()
         {
             List<SqlParserBaselineTestInput> inputs = new List<SqlParserBaselineTestInput>()
             {
                 // Positive
-                CreateInput(description: "Basic", offsetLimitClause: "OFFSET 10 LIMIT 10"),
-                CreateInput(description: "Parameters", offsetLimitClause: "OFFSET @OFFSETCOUNT LIMIT @LIMITCOUNT"),
+                CreateInput(description: "Basic", orderByClause: "ORDER BY 1"),
+                CreateInput(description: "Ascending", orderByClause: "ORDER BY 1 asc"),
+                CreateInput(description: "Multi Item Array", orderByClause: "ORDER BY 1 DESC"),
+                CreateInput(description: "Case Insensitive", orderByClause: "OrDeR By 1 DeSc"),
 
                 // Negative
-                CreateInput(description: "Non integer or paramter count", offsetLimitClause: "OFFSET 'asdf' LIMIT 10"),
-                CreateInput(description: "Offset without limit", offsetLimitClause: "OFFSET 10"),
-                CreateInput(description: "Limit without offset", offsetLimitClause: "LIMIT 10"),
+                CreateInput(description: "No spaces", orderByClause: "ORDERBY 1"),
+                CreateInput(description: "Wrong Keyword", orderByClause: "ORER BY 1"),
             };
 
             this.ExecuteTestSuite(inputs);
         }
 
-        public static SqlParserBaselineTestInput CreateInput(string description, string offsetLimitClause)
+        [TestMethod]
+        public void MultiOrderBy()
         {
-            return new SqlParserBaselineTestInput(description, $"SELECT * {offsetLimitClause}");
+            List<SqlParserBaselineTestInput> inputs = new List<SqlParserBaselineTestInput>()
+            {
+                // Positive
+                CreateInput(description: "Basic", orderByClause: "ORDER BY 1, 2, 3"),
+                CreateInput(description: "Only one sort order", orderByClause: "ORDER BY 1, 2 DESC, 3"),
+                CreateInput(description: "All sort order", orderByClause: "ORDER BY 1 ASC, 2 DESC, 3 ASC"),
+
+                // Negative
+                CreateInput(description: "Trailing comma", orderByClause: "ORDER BY 1 ASC,"),
+            };
+
+            this.ExecuteTestSuite(inputs);
+        }
+
+        public static SqlParserBaselineTestInput CreateInput(string description, string orderByClause)
+        {
+            return new SqlParserBaselineTestInput(description, $"SELECT * {orderByClause}");
         }
     }
 }
